HA improvements (#764)
* AF-2069: After space is deleted other users are not redirected to the spaces view on Openshift [CAT1]

* AF-2060: Dialog about deleted asset is opened several times on OpenShift [CAT1]

* AF-2052: No error message when assets are imported on the same pod (OpenShift)[Cat1]

* Unit test

+++ uberfire-server/src/main/java/org/uberfire/server/FileUploadServlet.java
@@ -40,6 +40,7 @@ public class FileUploadServlet
 
     private static final String RESPONSE_OK = "OK";
     private static final String RESPONSE_FAIL = "FAIL";
+    private static final String RESPONSE_CONFLICT = "CONFLICT";
 
     @Inject
     @Named("ioStrategy")
@@ -94,9 +95,22 @@ public class FileUploadServlet
 
         final Path path = ioService.get(uri);
 
-        writeFile(ioService,
-                  path,
-                  fileItem);
+        try {
+            ioService.startBatch(path.getFileSystem());
+
+            if (ioService.exists(path)) {
+                writeResponse(response,
+                              RESPONSE_CONFLICT);
+                response.sendError(HttpServletResponse.SC_CONFLICT);
+                return;
+            }
+
+            writeFile(ioService,
+                      path,
+                      fileItem);
+        } finally {
+            ioService.endBatch();
+        }
 
         writeResponse(response,
                       RESPONSE_OK);
